[server] Tests housekeeping and improved robustness
* There is now a new common base class for QGIS Server testing
* WMS tests have been moved into their own class
* Added tests for onlineResource URL in query string
