+++ tests/src/python/CMakeLists.txt
@@ -189,8 +189,9 @@ IF (WITH_APIDOC)
 ENDIF (WITH_APIDOC)
 
 IF (WITH_SERVER)
-ADD_PYTHON_TEST(PyQgsServer test_qgsserver.py)
-ADD_PYTHON_TEST(PyQgsServerPlugins test_qgsserver_plugins.py)
+  ADD_PYTHON_TEST(PyQgsServer test_qgsserver.py)
+  ADD_PYTHON_TEST(PyQgsServerPlugins test_qgsserver_plugins.py)
+  ADD_PYTHON_TEST(PyQgsServerWMS test_qgsserver_wms.py)
   ADD_PYTHON_TEST(PyQgsServerSettings test_qgsserver_settings.py)
   ADD_PYTHON_TEST(PyQgsServerProjectUtils test_qgsserver_projectutils.py)
   ADD_PYTHON_TEST(PyQgsServerSecurity test_qgsserver_security.py)
