+++ tests/testdata/qgis_server/wfs_describefeaturetype.txt
@@ -1,14 +1,14 @@
 Content-Length: 913
 Content-Type: text/xml; charset=utf-8
 
-<schema xmlns:ogc="http://www.opengis.net/ogc" elementFormDefault="qualified" targetNamespace="http://www.qgis.org/gml" version="1.0" xmlns="http://www.w3.org/2001/XMLSchema" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:gml="http://www.opengis.net/gml" xmlns:qgs="http://www.qgis.org/gml">
+<schema xmlns:gml="http://www.opengis.net/gml" targetNamespace="http://www.qgis.org/gml" xmlns:qgs="http://www.qgis.org/gml" xmlns="http://www.w3.org/2001/XMLSchema" xmlns:ogc="http://www.opengis.net/ogc" version="1.0" xmlns:xsd="http://www.w3.org/2001/XMLSchema" elementFormDefault="qualified">
  <import namespace="http://www.opengis.net/gml" schemaLocation="http://schemas.opengis.net/gml/2.1.2/feature.xsd"/>
  <element type="qgs:testlayerType" name="testlayer" substitutionGroup="gml:_Feature"/>
  <complexType name="testlayerType">
   <complexContent>
    <extension base="gml:AbstractFeatureType">
     <sequence>
-     <element minOccurs="0" maxOccurs="1" type="gml:GeometryPropertyType" name="geometry"/>
+     <element maxOccurs="1" type="gml:GeometryPropertyType" minOccurs="0" name="geometry"/>
      <element type="long" name="id"/>
      <element type="string" name="name"/>
      <element type="string" name="utf8nameè"/>
