+++ tests/testdata/qgis_server/wfs_getcapabilities.txt
@@ -1,11 +1,11 @@
-Content-Length: 2781
+Content-Length: 3001
 Content-Type: text/xml; charset=utf-8
 
-<WFS_Capabilities xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:ogc="http://www.opengis.net/ogc" version="1.0.0" xmlns="http://www.opengis.net/wfs" updateSequence="0" xmlns:xlink="http://www.w3.org/1999/xlink" xsi:schemaLocation="http://www.opengis.net/wfs http://schemas.opengis.net/wfs/1.0.0/WFS-capabilities.xsd" xmlns:gml="http://www.opengis.net/gml" xmlns:ows="http://www.opengis.net/ows">
+<WFS_Capabilities xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:gml="http://www.opengis.net/gml" xmlns:ows="http://www.opengis.net/ows" xmlns="http://www.opengis.net/wfs" xsi:schemaLocation="http://www.opengis.net/wfs http://schemas.opengis.net/wfs/1.0.0/WFS-capabilities.xsd" xmlns:ogc="http://www.opengis.net/ogc" version="1.0.0" xmlns:xlink="http://www.w3.org/1999/xlink" updateSequence="0">
  <Service>
   <Name>WFS</Name>
   <Title>QGIS TestProject</Title>
-  <Abstract>Some UTF8 text èòù</Abstract>
+  <Abstract><![CDATA[Some UTF8 text èòù]]></Abstract>
   <OnlineResource/>
  </Service>
  <Capability>
@@ -13,12 +13,12 @@ Content-Type: text/xml; charset=utf-8
    <GetCapabilities>
     <DCPType>
      <HTTP>
-      <Get onlineResource="?"/>
+      <Get onlineResource="?MAP=tests/testdata/qgis_server/test_project_wfs.qgs"/>
      </HTTP>
     </DCPType>
     <DCPType>
      <HTTP>
-      <Post onlineResource="?"/>
+      <Post onlineResource="?MAP=tests/testdata/qgis_server/test_project_wfs.qgs"/>
      </HTTP>
     </DCPType>
    </GetCapabilities>
@@ -28,12 +28,12 @@ Content-Type: text/xml; charset=utf-8
     </SchemaDescriptionLanguage>
     <DCPType>
      <HTTP>
-      <Get onlineResource="?"/>
+      <Get onlineResource="?MAP=tests/testdata/qgis_server/test_project_wfs.qgs"/>
      </HTTP>
     </DCPType>
     <DCPType>
      <HTTP>
-      <Post onlineResource="?"/>
+      <Post onlineResource="?MAP=tests/testdata/qgis_server/test_project_wfs.qgs"/>
      </HTTP>
     </DCPType>
    </DescribeFeatureType>
@@ -45,19 +45,19 @@ Content-Type: text/xml; charset=utf-8
     </ResultFormat>
     <DCPType>
      <HTTP>
-      <Get onlineResource="?"/>
+      <Get onlineResource="?MAP=tests/testdata/qgis_server/test_project_wfs.qgs"/>
      </HTTP>
     </DCPType>
     <DCPType>
      <HTTP>
-      <Post onlineResource="?"/>
+      <Post onlineResource="?MAP=tests/testdata/qgis_server/test_project_wfs.qgs"/>
      </HTTP>
     </DCPType>
    </GetFeature>
    <Transaction>
     <DCPType>
      <HTTP>
-      <Post onlineResource="?"/>
+      <Post onlineResource="?MAP=tests/testdata/qgis_server/test_project_wfs.qgs"/>
      </HTTP>
     </DCPType>
    </Transaction>
@@ -72,7 +72,7 @@ Content-Type: text/xml; charset=utf-8
    <Title>A test vector layer</Title>
    <Abstract>A test vector layer with unicode òà</Abstract>
    <SRS>EPSG:4326</SRS>
-   <LatLongBoundingBox maxx="8.20355" minx="8.20346" maxy="44.9015" miny="44.9014"/>
+   <LatLongBoundingBox maxy="44.9015" maxx="8.20355" miny="44.9014" minx="8.20346"/>
    <Operations>
     <Query/>
     <Insert/>
