+++ tests/testdata/qgis_server/wfs_getfeature_startindex2_post.txt
@@ -2,41 +2,24 @@ Content-Type: text/xml; charset=utf-8
 
 <wfs:FeatureCollection xmlns:wfs="http://www.opengis.net/wfs" xmlns:ogc="http://www.opengis.net/ogc" xmlns:gml="http://www.opengis.net/gml" xmlns:ows="http://www.opengis.net/ows" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:qgs="http://www.qgis.org/gml" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.opengis.net/wfs http://schemas.opengis.net/wfs/1.0.0/wfs.xsd http://www.qgis.org/gml ?MAP=tests/testdata/qgis_server/test_project_wfs.qgs&amp;SERVICE=WFS&amp;VERSION=1.0.0&amp;REQUEST=DescribeFeatureType&amp;TYPENAME=testlayer&amp;OUTPUTFORMAT=XMLSCHEMA"><gml:boundedBy>
  <gml:Box srsName="EPSG:4326">
-  <gml:coordinates cs="," ts=" ">8.20345931,44.90139484 8.20354699,44.90148253</gml:coordinates>
+  <gml:coordinates cs="," ts=" ">8,44 9,45</gml:coordinates>
  </gml:Box>
 </gml:boundedBy>
 <gml:featureMember>
- <qgs:testlayer fid="testlayer.0">
+ <qgs:testlayer fid="testlayer.2">
   <gml:boundedBy>
    <gml:Box srsName="EPSG:4326">
-    <gml:coordinates cs="," ts=" ">8.20349634,44.90148253 8.20349634,44.90148253</gml:coordinates>
+    <gml:coordinates cs="," ts=" ">8.20345931,44.90139484 8.20345931,44.90139484</gml:coordinates>
    </gml:Box>
   </gml:boundedBy>
   <qgs:geometry>
    <Point xmlns="http://www.opengis.net/gml" srsName="EPSG:4326">
-    <coordinates xmlns="http://www.opengis.net/gml">8.20349634,44.90148253</coordinates>
+    <coordinates xmlns="http://www.opengis.net/gml">8.20345931,44.90139484</coordinates>
    </Point>
   </qgs:geometry>
-  <qgs:id>1</qgs:id>
-  <qgs:name>one</qgs:name>
-  <qgs:utf8nameè>one èé</qgs:utf8nameè>
- </qgs:testlayer>
-</gml:featureMember>
-<gml:featureMember>
- <qgs:testlayer fid="testlayer.1">
-  <gml:boundedBy>
-   <gml:Box srsName="EPSG:4326">
-    <gml:coordinates cs="," ts=" ">8.20354699,44.90143568 8.20354699,44.90143568</gml:coordinates>
-   </gml:Box>
-  </gml:boundedBy>
-  <qgs:geometry>
-   <Point xmlns="http://www.opengis.net/gml" srsName="EPSG:4326">
-    <coordinates xmlns="http://www.opengis.net/gml">8.20354699,44.90143568</coordinates>
-   </Point>
-  </qgs:geometry>
-  <qgs:id>2</qgs:id>
-  <qgs:name>two</qgs:name>
-  <qgs:utf8nameè>two àò</qgs:utf8nameè>
+  <qgs:id>3</qgs:id>
+  <qgs:name>three</qgs:name>
+  <qgs:utf8nameè>three èé↓</qgs:utf8nameè>
  </qgs:testlayer>
 </gml:featureMember>
 </wfs:FeatureCollection>
