+++ tests/testdata/qgis_server/wms_getfeatureinfo-text-html.txt
@@ -1,4 +1,4 @@
-Content-Length: 360
+Content-Length: 512
 Content-Type: text/html; charset=utf-8
 
 <HEAD>
